requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
aioredis==2.0.1
discord.py==2.3.2
aiohttp==3.9.0
pyyaml==6.0.1
python-dateutil==2.8.2
